Extract Header component
